Handle request/leave full screen events (partial implementation)
